add type and move to test file
